Move sw code to production
